Clean difference logging
