+++ pass-grant-loader/pass-grant-data/src/main/resources/logback.xml
@@ -26,7 +26,7 @@
   <logger name="org.eclipse.pass" additivity="false">
     <appender-ref ref="STDOUT"/>
   </logger>
-  <logger name="org.eclipse.pass.support.grant.data.JhuPassInitUpdater" additivity="false" level="INFO">
+  <logger name="org.eclipse.pass.support.grant.data.DifferenceLogger" additivity="false" level="INFO">
     <appender-ref ref="STDOUT"/>
   </logger>
 </configuration>
